chore(maintainability): reduce duplication and improve maintainability - reduce duplication - improve maintainability
[Delivers #161473724]
