+++ authors/apps/authentication/mixins.py
@@ -1,0 +1,15 @@
+from .renderers import UserJSONRenderer
+from rest_framework.permissions import(
+ AllowAny,
+ IsAuthenticated
+)
+
+
+class AttributeMixins:
+    permission_classes = (AllowAny,)
+    renderer_classes = (UserJSONRenderer,)
+
+
+class AuthenticatedUserMixins:
+    permission_classes = (IsAuthenticated,)
+    renderer_classes = (UserJSONRenderer,)
