+++ authors/apps/authentication/tests/test_social_login.py
